fixed rest controlles

+++ src/main/java/com/snoopgame/application/controllers/EmployeeController.java
@@ -1,15 +1,12 @@
 package com.snoopgame.application.controllers;
 
 import com.snoopgame.application.Entities.Employee;
-import com.snoopgame.application.Entities.Role;
-import com.snoopgame.application.Entities.User;
+import com.snoopgame.application.Entities.Order;
+import com.snoopgame.application.Entities.Status;
 import com.snoopgame.application.Repositories.EmployeeRepository;
-import com.snoopgame.application.objectsForJSON.Employees;
+import com.snoopgame.application.Repositories.OrderRepository;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -18,28 +15,40 @@ import java.util.Map;
 @Controller
 public class EmployeeController {
     private final EmployeeRepository employeeRepository;
+    private final OrderRepository orderRepository;
 
-    public EmployeeController(EmployeeRepository employeeRepository) {
+    public EmployeeController(EmployeeRepository employeeRepository, OrderRepository orderRepository) {
         this.employeeRepository = employeeRepository;
+        this.orderRepository = orderRepository;
     }
 
     @PostMapping("employee/add")
-    public String addEmployee(Employee employee, Map<String,Object> model){
-        Employee employeeFromDb = employeeRepository.findByNameAndSurnameAndMiddleName(employee.getName(),employee.getSurname(),employee.getMiddleName());
+    public String addEmployee(Employee employee, Map<String, Object> model) {
+        Employee employeeFromDb = employeeRepository.findByNameAndSurnameAndMiddleName(employee.getName(),
+                employee.getSurname(), employee.getMiddleName());
         if (employeeFromDb != null) {
             model.put("message", "Employee already exists!");
-            return "redirect:/main";
+            return "employee/add";
         }
         employeeRepository.save(employee);
         return "redirect:/main";
     }
+
     @PostMapping("employee/remove")
-    public String removeEmployee(Employee employee, Map<String,Object> model){
-        Employee employeeFromDb = employeeRepository.findByNameAndSurnameAndMiddleName(employee.getName(),employee.getSurname(),employee.getMiddleName());
+    public String removeEmployee(Employee employee, Map<String, Object> model) {
+        Employee employeeFromDb = employeeRepository.findByNameAndSurnameAndMiddleName(employee.getName(),
+                employee.getSurname(), employee.getMiddleName());
         if (employeeFromDb == null) {
             model.put("message", "Employee not exists!");
-            return "redirect:/main";
+            return "employee/remove";
         }
+        ArrayList<Order> orders = (ArrayList<Order>) orderRepository.findByEmployeeAndStatuses(employeeFromDb,
+                Collections.singleton(Status.INITIATED));
+        if (orders.size() != 0) {
+            model.put("message", "Employee don't returned all phones!");
+            return "employee/remove";
+        }
+        orderRepository.deleteAll(orderRepository.findByEmployee(employeeFromDb));
         employeeRepository.delete(employeeFromDb);
         return "redirect:/main";
     }
